requirements.txt
numpy>=1.22.2
torchvision>=0.11.3
path>=16.4.0
scipy>=1.8.0